tidy: split TestRunner to cc
