+++ .clang-tidy
@@ -4,6 +4,7 @@ Checks:          'abseil-*,
                   clang-diagnostic-*,
                   misc-unused-using-decls,
                   modernize-*,
+                  -modernize-use-trailing-return-type,
                   performance-*,
                   readability-braces-around-statements,
                   readability-container-size-empty,
